Update P1Q2_Reducer.py
Update, as existing code was correct, but if and nested if were in wrong order, loop and job ran, but output produce was blank file.  After I figured out the problem, correcting the order of ifs, loop and job still ran, but not output produced and was present in the resulting file.